chore: move NumberInputNode to another file
