+++ domain/node/input_node.go
@@ -14,20 +14,4 @@ type (
 		Node
 		InputType() InputType
 	}
-
-	NumberInputNode struct {
-		NodeID domain.ID
-
-		Label domain.String
-
-		Required bool
-
-		Unit *domain.String
-	}
 )
-
-var _ InputNode = (*NumberInputNode)(nil)
-
-func (n *NumberInputNode) ID() domain.ID        { return n.NodeID }
-func (n *NumberInputNode) Type() Type           { return NodeTypeInput }
-func (n *NumberInputNode) InputType() InputType { return InputTypeNumber }
